Correct retrieval of "token" flag value
When a caller supplies a nonempty value for the "token" flag, function
getGitHubToken would mistakenly look up the value of another (usually)
absent flag named by the "token" flag's value, instead of returning the
"token" flag's value itself.

That is, given an invocation like

  label-all-the-things --token xxx create ...

then instead of using "xxx" as the GitHub API token, the program would
instead look for the value of a flag named "xxx".